[Refactor] 댓글 model & dto 분리
