+++ TwoHoSun/DTO/Comment/CommentResponseModel.swift
@@ -1,5 +1,5 @@
 //
-//  CommentsModel.swift
+//  CommentResponseModel.swift
 //  TwoHoSun
 //
 //  Created by 235 on 10/22/23.
@@ -7,12 +7,12 @@
 
 import Foundation
 
-struct CommentsModel: Codable {
+struct CommentResponseObject: Codable {
     let commentId: Int
     let createDate: String
     let modifiedDate: String
     let content: String
-    let subComments: [CommentsModel]?
+    let subComments: [CommentResponseObject]?
     let isMine: Bool
-    let author: AuthorModel?
+    let author: AuthorResponseObject?
 }
