--- TwoHoSun/Model/Prev/CommonModel/CommonModel.swift
@@ -1,9 +1,0 @@
-//
-//  CommonModel.swift
-//  TwoHoSun
-//
-//  Created by 235 on 10/22/23.
-//
-
-import Foundation
-// MARK: 임시모델 공통된 모델들 여기에 놓으면 됩니다.
