--- TwoHoSun/Model/Prev/FeedModel/CommentPostModel.swift
@@ -1,12 +1,0 @@
-//
-//  CommentPostModel.swift
-//  TwoHoSun
-//
-//  Created by 235 on 10/23/23.
-//
-
-import Foundation
-struct CommentPostModel: Codable {
-    let postId: Int
-    let contents: String
-}
